requirements.txt
PyFunctional
beautifulsoup4
click
requests
dacite
dateparser